# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: GigShield API
  description: Parametric Income Protection for Q-Commerce Delivery Partners
  version: 2.1.0
paths:
...
  /api/triggers/check/{zone}:
    post:
      tags:
      - Triggers
      summary: Check Zone Triggers
      description: Check all 5 triggers for a zone (simulates 15-min polling cycle).
      operationId: check_zone_triggers_api_triggers_check__zone__post
      parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
          title: Zone
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/triggers/simulate/{trigger_type}:
    post:
      tags:
      - Triggers
      summary: Simulate Trigger
      description: Simulate a trigger event for demo — fires trigger and processes
        claims.
      operationId: simulate_trigger_api_triggers_simulate__trigger_type__post
      security:
      - HTTPBearer: []
      parameters:
      - name: trigger_type
        in: path
        required: true
        schema:
          type: string
          title: Trigger Type
      - name: zone
        in: query
        required: false
        schema:
          type: string
          default: Indiranagar
          title: Zone
      - name: severity
        in: query
        required: false
        schema:
          type: string
          default: severe
          title: Severity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/triggers/history:
    get:
      tags:
      - Triggers
      summary: Get Trigger History
      description: Get trigger history.
      operationId: get_trigger_history_api_triggers_history_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/triggers/{trigger_id}/resolve:
    put:
      tags:
      - Triggers
      summary: Resolve Trigger
      description: Mark a trigger as resolved.
      operationId: resolve_trigger_api_triggers__trigger_id__resolve_put
      security:
      - HTTPBearer: []
      parameters:
      - name: trigger_id
        in: path
        required: true
        schema:
          type: integer
          title: Trigger Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/triggers/config:
    get:
      tags:
      - Triggers
      summary: Get Config
      description: Get trigger configuration (thresholds, logic gates).
      operationId: get_config_api_triggers_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/triggers/zone-status/{zone}:
    get:
      tags:
      - Triggers
      summary: Get Zone Live Status
      description: Get live status of all environmental signals for a zone.
      operationId: get_zone_live_status_api_triggers_zone_status__zone__get
      parameters:
      - name: zone
        in: path
        required: true
        schema:
          type: string
          title: Zone
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer